+++ tests/integration/butterfree/core/transform/test_feature_set.py
@@ -55,7 +55,13 @@ class TestFeatureSet:
                     ),
                 ),
             ],
-            keys=[KeyFeature(name="id", description="The user's Main ID or device ID")],
+            keys=[
+                KeyFeature(
+                    name="id",
+                    description="The user's Main ID or device ID",
+                    dtype=DataType.INTEGER,
+                )
+            ],
             timestamp=TimestampFeature(),
         )
